literacy_app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-CORS
PyJWT # For JWT token generation
gunicorn # Common WSGI server, good practice for PythonAnywhere
